Add docker-compose with neo4j service
This adds docker-compose with neo4j basic setup. It creates volumes in project directory for neo4j logs, config, data and plugins, so that data is not lost after container is down. It also adds README with docker and neo4j "how-to" and .gitignore file with neo4j volume directory